+++ components/organisms/navigation/CommonNav.tsx
@@ -1,5 +1,6 @@
 "use client";
 import { useState } from "react";
+import { usePathname } from "next/navigation";
 
 import { Menu } from "@/types/Menu";
 import styles from "./CommonNav.module.css";
@@ -9,6 +10,8 @@ import navJson from "@/data/nav.json";
 function CommonNav() {
   const [menus] = useState<Menu[]>(navJson);
 
+  const path = usePathname();
+
   return (
     <nav className={styles.navigation}>
       <div className={styles.menu}>
@@ -16,7 +19,9 @@ function CommonNav() {
           <Link
             href={menu.path}
             key={menu.id}
-            className="text-white hover:text-gray-500"
+            className={`text-white hover:text-gray-500 ${
+              path.startsWith(menu.path) ? "bg-blue-700 text-white" : ""
+            }`}
           >
             <span className={styles.label}>{menu.label}</span>
           </Link>